+++ jsplat-io-sog/src/main/java/de/javagl/jsplat/io/sog/Clustering.java
@@ -46,7 +46,7 @@ class Clustering
         /**
          * The centroids
          */
-        float centroids[][];
+        double centroids[][];
         
         /**
          * The labels
